Moved site notice generation to separate function
