+++ app/page.tsx
@@ -1,28 +1,15 @@
 import Link from "next/link";
 import IPage from "./models/IPage";
 import ISubpage from "./models/ISubpage";
-import { getPages } from "@/app/utils/pages";
+import { getPages, getSiteNotice } from "@/app/utils/pages";
 import "./styles/homepage.scss";
 
 export default async function Home() {
     const pages = await getPages();
-    const activeLeagues = pages.filter(page => page.name.indexOf("Current (") > -1);
-    let activeLeaguesAsString = "";
-    if(activeLeagues.length === 1){
-        activeLeaguesAsString = activeLeagues[0].showAndSeason
-    } else if (activeLeagues.length === 2) {
-        activeLeaguesAsString = `${activeLeagues[0].showAndSeason} and ${activeLeagues[1].showAndSeason}`
-    } else if(activeLeagues.length > 2){
-        const activeLeaguesWithoutLast = activeLeagues.toSpliced(activeLeagues.length - 2, 1).map(league => league.showAndSeason);
-        activeLeaguesAsString = `${activeLeaguesWithoutLast.join(", ")}, and ${activeLeagues[activeLeagues.length - 1].showAndSeason}`
-    }
+    const siteNotice = getSiteNotice(pages);
     return (
         <>
-            <div className="site-notice">
-                <p>
-                    The league{activeLeagues.length > 1 && "s"} for {activeLeaguesAsString} is already underway. Stay tuned to see how you fair. If you would still like to join this or future leagues, feel free to email <Link className="standard-link" href="mailto:xfactorleaguesite@gmail.com">xfactorleaguesite@gmail.com</Link>.
-                </p>
-            </div>
+            {siteNotice}
             {pages.map((p: IPage) => { 
                 const keyName = p.name.toLowerCase().replaceAll(" ", "-");
                 return (<div key={`links-section-${keyName}`}>
