+++ app/utils/pages.tsx
@@ -1,4 +1,5 @@
 import { getLeagueConfigurationKeys, hasContestantData } from "../dataSources/dbFetch";
+import Link from "next/link";
 
 interface ILeagueLink {
     name: string
@@ -93,6 +94,24 @@ export async function getPages(): Promise<ILeagueLink[]> {
     }
     const paths:Array<ILeagueLink> = [...activeLeaguePaths,...archiveLeaguePaths];
     return paths;
+}
+
+export function getSiteNotice(pages:ILeagueLink[]){
+    const activeLeagues = pages.filter(page => page.name.indexOf("Current (") > -1);
+    let activeLeaguesAsString = "";
+    if(activeLeagues.length === 1){
+        activeLeaguesAsString = activeLeagues[0].showAndSeason
+    } else if (activeLeagues.length === 2) {
+        activeLeaguesAsString = `${activeLeagues[0].showAndSeason} and ${activeLeagues[1].showAndSeason}`
+    } else if(activeLeagues.length > 2){
+        const activeLeaguesWithoutLast = activeLeagues.toSpliced(activeLeagues.length - 2, 1).map(league => league.showAndSeason);
+        activeLeaguesAsString = `${activeLeaguesWithoutLast.join(", ")}, and ${activeLeagues[activeLeagues.length - 1].showAndSeason}`
+    }
+    return <div className="site-notice">
+        <p>
+                    The league{activeLeagues.length > 1 && "s"} for {activeLeaguesAsString} is already underway. Stay tuned to see how you fair. If you would still like to join this or future leagues, feel free to email <Link className="standard-link" href="mailto:xfactorleaguesite@gmail.com">xfactorleaguesite@gmail.com</Link>.
+        </p>
+    </div>
 }
 
 export function getUrlParams (dataKeys:string[]){
